third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/get-text.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/text-level-semantics.html#dom-a-text">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that HTMLAnchorElement.text returns the same value as the textContent attribute.");

var div = document.createElement('div');
div.innerHTML = '<a>a<br>b</a>';
var a = div.firstChild;
shouldBe("a.__proto__", "HTMLAnchorElement.prototype");

shouldBeEqualToString('a.text', 'ab');
shouldBeEqualToString('a.textContent', 'ab');
</script>
</body>
</html>
