third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/get-text-expected.txt

Tests that HTMLAnchorElement.text returns the same value as the textContent attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS a.__proto__ is HTMLAnchorElement.prototype
PASS a.text is "ab"
PASS a.textContent is "ab"
PASS successfullyParsed is true

TEST COMPLETE
